requirements.txt
streamlit==1.37.1
pandas==2.2.2
gspread==6.1.2
google-auth==2.33.0
xlsxwriter==3.2.0
